+++ ExilenceNextApp/src/components/custom-price-dialog/CustomPriceDialog.styles.ts
@@ -1,0 +1,9 @@
+import { makeStyles } from '@material-ui/core';
+
+const useStyles = makeStyles((theme) => ({
+  dialogActions: {
+    padding: theme.spacing(2),
+  },
+}));
+
+export default useStyles;