+++ ExilenceNextApp/src/components/price-league-dropdown/PriceLeagueDropdown.styles.ts
@@ -1,0 +1,9 @@
+import { makeStyles } from '@material-ui/core';
+
+const useStyles = makeStyles(() => ({
+  small: {
+    fontSize: '0.875rem',
+  },
+}));
+
+export default useStyles;