+++ ExilenceNextApp/src/store/uiStateStore.ts
@@ -3,6 +3,7 @@ import { action, makeObservable, observable, runInAction } from 'mobx';
 import { persist } from 'mobx-persist';
 import { map } from 'rxjs/operators';
 import { v4 as uuidv4 } from 'uuid';
+import { IPricedItem } from '../interfaces/priced-item.interface';
 
 import { IStashTab } from '../interfaces/stash.interface';
 import { IStatusMessage } from '../interfaces/status-message.interface';
@@ -18,7 +19,7 @@ export type GroupDialogType = 'create' | 'join' | undefined;
 export class UiStateStore {
   @observable @persist userId: string = uuidv4();
   @observable sessIdCookie: ICookie | undefined = undefined;
-  @persist @observable sidenavOpen: boolean = true;
+  @persist @observable sidenavOpen: boolean = false;
   @persist @observable toolbarTourOpen: boolean = true;
   @observable validated: boolean = false;
   @observable isValidating: boolean = false;
@@ -30,6 +31,7 @@ export class UiStateStore {
   @observable notificationList: Notification[] = [];
   @observable initiated: boolean = false;
   @observable itemTableFilterText: string = '';
+  @observable priceTableFilterText: string = '';
   @observable isInitiating: boolean = false;
   @observable groupDialogOpen: boolean = false;
   @observable groupDialogType: 'create' | 'join' | undefined = undefined;
@@ -46,6 +48,7 @@ export class UiStateStore {
   @observable leavingGroup: boolean = false;
   @observable clearingSnapshots: boolean = false;
   @observable profilesLoaded: boolean = false;
+  @observable settingsTabIndex: number = 0;
   @observable filteredStashTabs: IStashTab[] | undefined = undefined;
   @persist @observable showItemTableFilter: boolean = false;
   @observable changingProfile: boolean = false;
@@ -56,6 +59,9 @@ export class UiStateStore {
   @observable statusMessage: IStatusMessage | undefined = undefined;
   @observable loginError: string | undefined = undefined;
   @persist @observable chartTimeSpan: TimespanType = 'All time';
+  @observable customPriceDialogOpen: boolean = false;
+  @observable selectedPricedItem: IPricedItem | undefined = undefined;
+  @persist @observable selectedPriceTableLeagueId: string | undefined = undefined;
 
   constructor(private rootStore: RootStore) {
     makeObservable(this);
@@ -64,6 +70,16 @@ export class UiStateStore {
   @action
   resetStatusMessage() {
     this.statusMessage = undefined;
+  }
+
+  @action.bound
+  setSettingsTabIndex(index: number) {
+    this.settingsTabIndex = index;
+  }
+
+  @action
+  setSelectedPriceTableLeagueId(id: string) {
+    this.selectedPriceTableLeagueId = id;
   }
 
   @action
@@ -231,6 +247,17 @@ export class UiStateStore {
   }
 
   @action
+  setSelectedPricedItem(item?: IPricedItem) {
+    this.selectedPricedItem = item;
+  }
+
+  @action
+  setCustomPriceDialogOpen(open: boolean, row?: IPricedItem) {
+    this.customPriceDialogOpen = open;
+    this.setSelectedPricedItem(row);
+  }
+
+  @action
   setSessIdCookie(sessionId: string) {
     const cookie = constructCookie(sessionId);
     return authService.setAuthCookie(cookie).pipe(
@@ -277,6 +304,11 @@ export class UiStateStore {
   @action
   setItemTableFilterText(text: string) {
     this.itemTableFilterText = text;
+  }
+
+  @action
+  setPriceTableFilterText(text: string) {
+    this.priceTableFilterText = text;
   }
 
   @action