+++ ExilenceNextApp/src/utils/item.utils.ts
@@ -7,7 +7,6 @@ import { IProperty } from '../interfaces/property.interface';
 import { ISocket } from '../interfaces/socket.interface';
 import { IStashTabSnapshot } from '../interfaces/stash-tab-snapshot.interface';
 import { ICompactTab, IStashTab } from '../interfaces/stash.interface';
-import { ITableItem } from '../interfaces/table-item.interface';
 
 export function mergeItemStacks(items: IPricedItem[]) {
   const mergedItems: IPricedItem[] = [];
@@ -45,11 +44,8 @@ export function formatSnapshotsForTable(stashTabSnapshots: IStashTabSnapshot[]) 
   return mergeItemStacks(mergedStashTabs);
 }
 
-export function mapPricedItemToTableItem(pricedItem: IPricedItem) {
-  return {
-    ...pricedItem,
-    tabNames: pricedItem.tab ? pricedItem.tab.map((t) => t.n).join(', ') : '',
-  } as ITableItem;
+export function parseTabNames(tabs: ICompactTab[]) {
+  return tabs.map((t) => t.n).join(', ');
 }
 
 export function mapItemsToPricedItems(items: IItem[], tab?: IStashTab) {
@@ -115,6 +111,7 @@ export function findItem<T extends IPricedItem>(array: T[], itemToFind: T) {
       (x.frameType === itemToFind.frameType || (x.name.indexOf(' Map') > -1 && x.frameType !== 3))
   );
 }
+
 export function isDivinationCard(icon: string) {
   return icon.indexOf('/Divination/') > -1;
 }
@@ -140,6 +137,8 @@ const rarities: (keyof Rarity)[] = [
   'currency', //5
   'divination', //6
   'quest', //7
+  'unknown', //8
+  'legacy', //9
 ];
 
 export function getRarity(identifier: number): keyof Rarity {